# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FoodFlow API",
    "version": "1.0.0",
    "description": "API para o sistema FoodFlow, um SaaS para gestão de restaurantes. Inclui autenticação, gestão de restaurantes, usuários, mesas, cardápio, sessões, pedidos, relatórios, reservas e personalização de estilo."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    },
    {
      "url": "https://foodflow-api.render.com",
      "description": "Servidor de produção (Render)"
    }
  ],
...
  "paths": {
...
    "/reports/sales/{restaurantId}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Gera relatório de vendas",
        "description": "Gera um relatório de vendas para um restaurante, com filtros por data e formato (JSON, PDF, CSV). Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID do restaurante"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Data inicial do filtro (ISO 8601)"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Data final do filtro (ISO 8601)"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "pdf",
                "csv"
              ],
              "description": "Formato do relatório"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Relatório de vendas gerado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesReport"
                  }
                },
                "example": [
                  {
                    "order_id": "order_1",
                    "total": 59.80,
                    "date": "2025-07-19T12:00:00Z"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Erro ao gerar relatório",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Restaurante não encontrado"
                }
              }
            }
          }
        }
      }
    },
    "/reports/occupancy/{restaurantId}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Gera relatório de ocupação",
        "description": "Gera um relatório de ocupação de mesas para um restaurante, com filtros por data e formato (JSON, PDF, CSV). Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID do restaurante"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Data inicial do filtro (ISO 8601)"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Data final do filtro (ISO 8601)"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "pdf",
                "csv"
              ],
              "description": "Formato do relatório"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Relatório de ocupação gerado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OccupancyReport"
                  }
                },
                "example": [
                  {
                    "table_id": "table_1",
                    "status": "occupied",
                    "date": "2025-07-19T12:00:00Z"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Erro ao gerar relatório",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Restaurante não encontrado"
                }
              }
            }
          }
        }
      }
    },
    "/reports/inventory/{restaurantId}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Gera relatório de estoque",
        "description": "Gera um relatório de estoque para um restaurante, com filtros por data e formato (JSON, PDF, CSV). Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID do restaurante"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Data inicial do filtro (ISO 8601)"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Data final do filtro (ISO 8601)"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "pdf",
                "csv"
              ],
              "description": "Formato do relatório"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Relatório de estoque gerado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryReport"
                  }
                },
                "example": [
                  {
                    "item_id": "item_1",
                    "name": "Hambúrguer",
                    "quantity_used": 10,
                    "date": "2025-07-19T12:00:00Z"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Erro ao gerar relatório",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Restaurante não encontrado"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerauth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via /auth/login"
      }
    },
    "schemas": {
...
      "SalesReport": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do pedido"
          },
          "total": {
            "type": "number",
            "description": "Valor total do pedido"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Data do pedido"
          }
        },
        "required": [
          "order_id",
          "total",
          "date"
        ]
      },
      "OccupancyReport": {
        "type": "object",
        "properties": {
          "table_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID da mesa"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "occupied",
              "reserved"
            ],
            "description": "Status da mesa"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Data do registro"
          }
        },
        "required": [
          "table_id",
          "status",
          "date"
        ]
      },
      "InventoryReport": {
        "type": "object",
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do item"
          },
          "name": {
            "type": "string",
            "description": "Nome do item"
          },
          "quantity_used": {
            "type": "integer",
            "description": "Quantidade utilizada"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Data do registro"
          }
        },
        "required": [
          "item_id",
          "name",
          "quantity_used",
          "date"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensagem de erro",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "description": "Mensagem de erro específica"
                },
                "path": {
                  "type": "string",
                  "description": "Campo com erro"
                },
                "location": {
                  "type": "string",
                  "description": "Local do erro (ex.: body, query)"
                }
              }
            },
            "description": "Lista de erros detalhados",
            "nullable": true
          }
        }
      }
    }
  }
}